+++ apps/web/pages/auth/sso/direct.tsx
@@ -1,13 +1,16 @@
+"use client";
+
 import { signIn } from "next-auth/react";
 import { useRouter } from "next/navigation";
 import { useEffect } from "react";
 
-import { samlProductID, samlTenantID } from "@calcom/features/ee/sso/lib/saml";
 import { HOSTED_CAL_FEATURES } from "@calcom/lib/constants";
 
 import type { inferSSRProps } from "@lib/types/inferSSRProps";
 
 import PageWrapper from "@components/PageWrapper";
+
+import type { getServerSideProps } from "@server/lib/auth/sso/direct/getServerSideProps";
 
 // This page is used to initiate the SAML authentication flow by redirecting to the SAML provider.
 // Accessible only on self-hosted Cal.com instances.
@@ -36,13 +39,6 @@ export default function Page({ samlTenantID, samlProductID }: inferSSRProps<type
   return null;
 }
 
-Page.PageWrapper = PageWrapper;
+export { getServerSideProps };
 
-export async function getServerSideProps() {
-  return {
-    props: {
-      samlTenantID,
-      samlProductID,
-    },
-  };
-}
+Page.PageWrapper = PageWrapper;
