+++ apps/web/server/lib/auth/sso/direct/getServerSideProps.tsx
@@ -1,0 +1,10 @@
+import { samlProductID, samlTenantID } from "@calcom/features/ee/sso/lib/saml";
+
+export async function getServerSideProps() {
+  return {
+    props: {
+      samlTenantID,
+      samlProductID,
+    },
+  };
+}
